refactor(backend): split lib.rs into focused modules
Split 1800-line lib.rs into 4 focused modules:
- routes.rs (101 lines): route building + middleware setup
- handlers.rs (737 lines): core API handlers (files, tiles, publish)
- upload.rs (236 lines): file upload logic
- public.rs (466 lines): public tile API endpoints

lib.rs reduced to 240 lines (module declarations + re-exports + tests)

No behavior changes. All 62 integration tests pass.
