+++ backend/src/lib.rs
@@ -1,46 +1,20 @@
-use axum::{
-    extract::{DefaultBodyLimit, Multipart, Path as AxumPath, State},
-    http::{header, HeaderMap, StatusCode},
-    response::IntoResponse,
-    routing::{get, post},
-    Json, Router,
-};
-use axum_login::AuthManagerLayerBuilder;
-use chrono::Utc;
-use rand::RngCore;
-use std::path::{Path, PathBuf};
-use tokio::{
-    fs,
-    io::{AsyncWriteExt, BufWriter},
-};
-use tower_http::cors::CorsLayer;
-use tower_sessions::SessionManagerLayer;
-
 mod auth;
 mod auth_routes;
 mod config;
 mod db;
+mod handlers;
 mod http_errors;
 mod import;
 mod mbtiles;
 mod models;
 mod password;
+mod public;
+mod routes;
 mod session_store;
 mod test_routes;
 mod tiles;
+mod upload;
 mod validation;
-
-/// Type alias for file metadata from the database
-type FileMetadata = (
-    String,
-    Option<String>,
-    String,
-    Option<String>,
-    Option<String>,
-    Option<String>,
-    Option<i32>,
-    Option<i32>,
-);
 
 pub use auth::{AuthBackend, User};
 pub use auth_routes::build_auth_router;
@@ -49,1544 +23,15 @@ pub use db::{
     init_database, is_initialized, reconcile_processing_files, set_initialized, DEFAULT_DB_PATH,
     PROCESSING_RECONCILIATION_ERROR,
 };
-use duckdb::types::ValueRef;
-use http_errors::{bad_request, internal_error, payload_too_large};
-use import::import_spatial_data;
-use mbtiles::import_mbtiles;
+pub use handlers::validate_slug;
 pub use models::{
     AppState, ErrorResponse, FileItem, FileSchemaResponse, PreviewMeta, PublicTileMeta,
     PublicTileUrl, PublishRequest, PublishResponse,
 };
-use models::{FeaturePropertiesResponse, FeatureProperty};
 pub use password::{hash_password, validate_password_complexity, verify_password, PasswordError};
+pub use routes::{build_api_router, build_test_router};
 pub use session_store::DuckDBStore;
-use test_routes::add_test_routes;
-use tiles::build_mvt_select_sql;
 pub use validation::{validate_geojson, validate_shapefile_zip};
-
-pub fn build_api_router(state: AppState) -> Router {
-    build_api_router_with_auth(state, true)
-}
-
-pub fn build_test_router(state: AppState) -> Router {
-    build_api_router_with_auth(state, false)
-}
-
-fn build_api_router_with_auth(state: AppState, with_auth: bool) -> Router {
-    // Read allowed origins from environment or use defaults
-    let allowed_origins = config::read_cors_origins();
-
-    // Build CORS layer with specific origins
-    // Note: When using credentials, we cannot use wildcards for headers
-    let mut cors = CorsLayer::new()
-        .allow_methods([
-            axum::http::Method::GET,
-            axum::http::Method::POST,
-            axum::http::Method::DELETE,
-        ])
-        .allow_headers([
-            axum::http::header::CONTENT_TYPE,
-            axum::http::header::ACCEPT,
-            axum::http::header::AUTHORIZATION,
-        ])
-        .allow_credentials(true);
-
-    // Add each allowed origin
-    for origin in allowed_origins {
-        if let Ok(parsed) = origin.parse::<axum::http::HeaderValue>() {
-            cors = cors.allow_origin(parsed);
-        } else {
-            eprintln!("Warning: Failed to parse CORS origin '{}', skipping. Check CORS_ALLOWED_ORIGINS environment variable.", origin);
-        }
-    }
-
-    let session_layer = SessionManagerLayer::new(state.session_store.clone())
-        .with_secure(config::read_cookie_secure())
-        .with_same_site(tower_cookies::cookie::SameSite::Lax);
-
-    let auth_layer =
-        AuthManagerLayerBuilder::new(state.auth_backend.clone(), session_layer).build();
-
-    let auth_router = build_auth_router();
-    let public_router = Router::new()
-        .route("/health", get(health_check))
-        .route("/api/test/is-initialized", get(check_is_initialized))
-        .route("/tiles/{slug}/{z}/{x}/{y}", get(get_public_tile))
-        .route(
-            "/tiles/{slug}",
-            get(get_public_pmtiles).head(head_public_pmtiles),
-        )
-        .route("/tiles/{slug}/meta", get(get_public_tile_meta));
-
-    let mut api_router = Router::new()
-        .route("/api/files", get(list_files))
-        .route("/api/uploads", post(upload_file))
-        .route("/api/files/{id}/preview", get(get_preview_meta))
-        .route("/api/files/{id}/tiles/{z}/{x}/{y}", get(get_tile))
-        .route(
-            "/api/files/{id}/features/{fid}",
-            get(get_feature_properties),
-        )
-        .route("/api/files/{id}/schema", get(get_file_schema))
-        .route("/api/files/{id}/publish", post(publish_file))
-        .route("/api/files/{id}/unpublish", post(unpublish_file))
-        .route("/api/files/{id}/public-url", get(get_public_url));
-
-    // Add authentication middleware if required
-    if with_auth {
-        api_router = api_router.route_layer(axum_login::login_required!(crate::AuthBackend));
-    }
-
-    // Combine all routes
-    let router = auth_router
-        .merge(public_router)
-        .merge(api_router)
-        .merge(add_test_routes(Router::new()));
-
-    router
-        .layer(DefaultBodyLimit::disable())
-        .with_state(state)
-        .layer(auth_layer)
-        .layer(cors)
-}
-
-async fn list_files(State(state): State<AppState>) -> impl IntoResponse {
-    let conn = state.db.lock().await;
-    let mut stmt = conn
-        .prepare(
-            "SELECT f.id, f.name, f.type, f.size, f.uploaded_at, f.status, f.crs, f.path, f.table_name, f.error, f.is_public, pf.slug
-          FROM files f
-          LEFT JOIN published_files pf ON f.id = pf.file_id
-          ORDER BY f.uploaded_at DESC",
-        )
-        .unwrap();
-
-    let items: Vec<FileItem> = stmt
-        .query_map([], |row| {
-            let table_name: Option<String> = row.get(8)?;
-            let error: Option<String> = row.get(9)?;
-            let is_public: bool = row.get(10).unwrap_or(false);
-            let public_slug: Option<String> = row.get(11).ok();
-            Ok(FileItem {
-                id: row.get(0)?,
-                name: row.get(1)?,
-                file_type: row.get(2)?,
-                size: row.get(3)?,
-                uploaded_at: {
-                    let ts: chrono::NaiveDateTime = row.get(4)?;
-                    ts.and_utc().to_rfc3339()
-                },
-                status: row.get(5)?,
-                crs: row.get(6)?,
-                path: row.get(7)?,
-                table_name,
-                error,
-                is_public: Some(is_public),
-                public_slug,
-            })
-        })
-        .unwrap()
-        .collect::<Result<Vec<_>, _>>()
-        .unwrap();
-
-    drop(conn);
-    Json(items)
-}
-
-async fn get_preview_meta(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    // Check if file exists and get meta
-    let mut stmt = conn
-        .prepare("SELECT name, crs, status, table_name, tile_format, tile_bounds, minzoom, maxzoom FROM files WHERE id = ?")
-        .map_err(internal_error)?;
-
-    let meta: Option<FileMetadata> = stmt
-        .query_row(duckdb::params![id], |row| {
-            Ok((
-                row.get(0)?,
-                row.get(1)?,
-                row.get(2)?,
-                row.get(3)?,
-                row.get(4)?,
-                row.get(5)?,
-                row.get(6)?,
-                row.get(7)?,
-            ))
-        })
-        .ok();
-
-    let (name, crs, status, table_name, tile_format, tile_bounds, minzoom, maxzoom) = match meta {
-        Some(m) => m,
-        None => {
-            return Err((
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "File not found".to_string(),
-                }),
-            ))
-        }
-    };
-
-    if status != "ready" {
-        return Err((
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready for preview".to_string(),
-            }),
-        ));
-    }
-
-    // Calculate BBOX in WGS84
-    let bbox_values = if let Some(bounds_json) = tile_bounds {
-        // MBTiles: use pre-calculated bounds
-        serde_json::from_str::<[f64; 4]>(&bounds_json).ok()
-    } else if let Some(tbl) = table_name {
-        // Dynamic: calculate ST_Extent (existing logic)
-        // Note: If CRS is missing/null, we assume EPSG:4326 and always_xy=true (lon/lat) for simplicity
-
-        // Query for bbox components directly
-        let bbox_components_query = format!(
-            "SELECT ST_XMin(b), ST_YMin(b), ST_XMax(b), ST_YMax(b) FROM (
-                SELECT ST_Extent(ST_Transform(geom, '{}', 'EPSG:4326', always_xy := true)) as b
-                FROM \"{tbl}\"
-            )",
-            crs.as_deref().unwrap_or("EPSG:4326")
-        );
-
-        conn.query_row(&bbox_components_query, [], |row| {
-            let minx: Option<f64> = row.get(0).ok();
-            let miny: Option<f64> = row.get(1).ok();
-            let maxx: Option<f64> = row.get(2).ok();
-            let maxy: Option<f64> = row.get(3).ok();
-
-            if let (Some(x1), Some(y1), Some(x2), Some(y2)) = (minx, miny, maxx, maxy) {
-                Ok([x1, y1, x2, y2])
-            } else {
-                Ok([0.0, 0.0, 0.0, 0.0]) // Handle empty result
-            }
-        })
-        .ok()
-        .filter(|b| b != &[0.0, 0.0, 0.0, 0.0])
-    } else {
-        None
-    };
-
-    Ok(Json(PreviewMeta {
-        id,
-        name,
-        crs,
-        bbox: bbox_values,
-        tile_format,
-        minzoom,
-        maxzoom,
-    }))
-}
-
-async fn get_tile(
-    State(state): State<AppState>,
-    AxumPath((id, z, x, y)): AxumPath<(String, i32, i32, i32)>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    validate_tile_coords(z, x, y)?;
-
-    println!(
-        "Received tile request: id={}, z={}, x={}, y={}",
-        id, z, x, y
-    );
-    let conn = state.db.lock().await;
-
-    // Get file metadata including tile_format
-    let (crs, status, table_name, tile_format, file_path): (
-        Option<String>,
-        String,
-        Option<String>,
-        Option<String>,
-        String,
-    ) = conn
-        .query_row(
-            "SELECT crs, status, table_name, tile_format, path FROM files WHERE id = ?",
-            duckdb::params![id],
-            |row| {
-                Ok((
-                    row.get(0)?,
-                    row.get(1)?,
-                    row.get(2)?,
-                    row.get(3)?,
-                    row.get(4)?,
-                ))
-            },
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "File not found".to_string(),
-                }),
-            )
-        })?;
-
-    if status != "ready" {
-        return Err((
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready for preview".to_string(),
-            }),
-        ));
-    }
-
-    // MBTiles branch
-    if let Some(format) = tile_format {
-        let full_path = mbtiles::resolve_mbtiles_path(&file_path);
-        drop(conn); // Release lock before async operation
-        match mbtiles::get_tile_from_mbtiles(&full_path, z, x, y).await {
-            Ok(Some(data)) => {
-                let ct = match format.as_str() {
-                    "mvt" => "application/vnd.mapbox-vector-tile",
-                    "png" => "image/png",
-                    _ => "application/octet-stream",
-                };
-
-                // Check if data is gzip-compressed (MBTiles tiles are often gzipped)
-                // Gzip magic bytes: 0x1f 0x8b
-                let is_gzipped = data.starts_with(&[0x1f, 0x8b]);
-
-                if is_gzipped {
-                    return Ok((
-                        [
-                            (header::CONTENT_TYPE, ct),
-                            (header::CONTENT_ENCODING, "gzip"),
-                        ],
-                        data,
-                    )
-                        .into_response());
-                } else {
-                    return Ok(([(header::CONTENT_TYPE, ct)], data).into_response());
-                }
-            }
-            Ok(None) => {
-                // Tile doesn't exist but coordinates are valid → 204 No Content
-                return Ok(StatusCode::NO_CONTENT.into_response());
-            }
-            Err(e) => {
-                return Err(internal_error(format!("Failed to read MBTiles: {}", e)));
-            }
-        }
-    }
-
-    // Dynamic generation branch (existing logic)
-    let table_name = table_name.ok_or_else(|| {
-        (
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready for preview".to_string(),
-            }),
-        )
-    })?;
-
-    let source_crs = crs.as_deref().unwrap_or("EPSG:4326");
-
-    // 2. Generate MVT
-    // logic:
-    //  - filter by source_id
-    //  - ST_Transform(geom, source_crs, 'EPSG:3857', always_xy := true)
-    //  - ST_TileEnvelope(z, x, y) to get tile bounds in 3857
-    //  - ST_AsMVTGeom(geom_3857, tile_env) to clip/transform to tile coords
-    //  - ST_AsMVT(...) to encode
-
-    // 2a. Build property struct keys based on captured column metadata.
-    // We keep property keys as original names for UX.
-    // Note: We exclude fid + geom.
-    let select_sql =
-        build_mvt_select_sql(&conn, &id, &table_name, source_crs).map_err(internal_error)?;
-
-    println!("Executing SQL for tile z={z} x={x} y={y} id={id}");
-
-    // Params: z, x, y (for AsMVTGeom bounds), z, x, y (for intersects)
-    let mvt_blob: Option<Vec<u8>> =
-        match conn.query_row(&select_sql, duckdb::params![z, x, y, z, x, y], |row| {
-            row.get(0)
-        }) {
-            Ok(blob) => Some(blob),
-            Err(e) => {
-                eprintln!("Tile Error (z={z}, x={x}, y={y}): {:?}", e);
-                eprintln!("SQL that failed: {}", select_sql);
-                return Err(internal_error(format!("Tile generation failed: {}", e)));
-            }
-        };
-
-    println!(
-        "Tile Request: z={z}, x={x}, y={y}, Blob Size: {:?}",
-        mvt_blob.as_ref().map(|v| v.len())
-    );
-
-    match mvt_blob {
-        Some(blob) if !blob.is_empty() => Ok((
-            [(header::CONTENT_TYPE, "application/vnd.mapbox-vector-tile")],
-            blob,
-        )
-            .into_response()),
-        _ => {
-            // Return empty response or 204? Mapbox clients usually expect 200 with empty body or valid PBF.
-            // An empty blob is a valid MVT (empty).
-            Ok((
-                [(header::CONTENT_TYPE, "application/vnd.mapbox-vector-tile")],
-                Vec::new(),
-            )
-                .into_response())
-        }
-    }
-}
-
-async fn get_feature_properties(
-    State(state): State<AppState>,
-    AxumPath((id, fid)): AxumPath<(String, i64)>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let (status, table_name, tile_format): (String, Option<String>, Option<String>) = conn
-        .query_row(
-            "SELECT status, table_name, tile_format FROM files WHERE id = ?",
-            duckdb::params![id],
-            |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "File not found".to_string(),
-                }),
-            )
-        })?;
-
-    // MBTiles files don't support feature properties
-    if tile_format.is_some() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse {
-                error: "Feature properties not available for MBTiles files".to_string(),
-            }),
-        ));
-    }
-
-    let table_name = table_name.filter(|_| status == "ready").ok_or_else(|| {
-        (
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready for preview".to_string(),
-            }),
-        )
-    })?;
-
-    let mut cols_stmt = conn
-        .prepare(
-            "SELECT normalized_name, original_name\n         FROM dataset_columns\n         WHERE source_id = ?\n         ORDER BY ordinal",
-        )
-        .map_err(internal_error)?;
-
-    let cols_iter = cols_stmt
-        .query_map(duckdb::params![&id], |row| {
-            let normalized: String = row.get(0)?;
-            let original: String = row.get(1)?;
-            Ok((normalized, original))
-        })
-        .map_err(internal_error)?;
-
-    let mut columns: Vec<(String, String)> = Vec::new();
-    for c in cols_iter {
-        columns.push(c.map_err(internal_error)?);
-    }
-
-    // Build a projection that preserves ordering and uses safe identifiers.
-    let mut select_exprs: Vec<String> = Vec::with_capacity(columns.len());
-    for (normalized, _original) in &columns {
-        select_exprs.push(format!("\"{normalized}\""));
-    }
-
-    let sql = format!(
-        "SELECT {} FROM \"{}\" WHERE fid = ?",
-        select_exprs.join(", "),
-        table_name
-    );
-
-    let mut stmt = conn.prepare(&sql).map_err(internal_error)?;
-    let mut rows = stmt.query(duckdb::params![fid]).map_err(internal_error)?;
-
-    let Some(row) = rows.next().map_err(internal_error)? else {
-        return Err((
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse {
-                error: "Feature not found".to_string(),
-            }),
-        ));
-    };
-
-    let mut properties: Vec<FeatureProperty> = Vec::with_capacity(columns.len());
-    for (index, (_normalized, original)) in columns.iter().enumerate() {
-        let raw = match row.get_ref(index).map_err(internal_error)? {
-            ValueRef::Null => serde_json::Value::Null,
-            ValueRef::Boolean(v) => serde_json::Value::Bool(v),
-            ValueRef::TinyInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::SmallInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::Int(v) => serde_json::Value::Number(v.into()),
-            ValueRef::BigInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::UTinyInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::USmallInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::UInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::UBigInt(v) => serde_json::Value::Number(v.into()),
-            ValueRef::Float(v) => serde_json::Number::from_f64(v as f64)
-                .map(serde_json::Value::Number)
-                .unwrap_or(serde_json::Value::Null),
-            ValueRef::Double(v) => serde_json::Number::from_f64(v)
-                .map(serde_json::Value::Number)
-                .unwrap_or(serde_json::Value::Null),
-            ValueRef::Text(bytes) => {
-                serde_json::Value::String(String::from_utf8_lossy(bytes).to_string())
-            }
-            ValueRef::Blob(bytes) => serde_json::Value::String(format!("0x{}", hex::encode(bytes))),
-            other => serde_json::Value::String(format!("{other:?}")),
-        };
-        properties.push(FeatureProperty {
-            key: original.clone(),
-            value: raw,
-        });
-    }
-
-    Ok(Json(FeaturePropertiesResponse { fid, properties }))
-}
-
-async fn get_file_schema(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let (status, tile_format, file_path): (String, Option<String>, String) = conn
-        .query_row(
-            "SELECT status, tile_format, path FROM files WHERE id = ?",
-            duckdb::params![id],
-            |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "File not found".to_string(),
-                }),
-            )
-        })?;
-
-    if status != "ready" {
-        drop(conn);
-        return Err((
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready".to_string(),
-            }),
-        ));
-    }
-
-    // Handle MBTiles files
-    if let Some(format) = tile_format {
-        drop(conn);
-        if format == "mvt" {
-            let full_path = mbtiles::resolve_mbtiles_path(&file_path);
-            match mbtiles::extract_mbtiles_layers(&full_path) {
-                Ok(layers) => {
-                    return Ok(Json(models::FileSchemaResponse { layers }));
-                }
-                Err(e) => {
-                    eprintln!("Failed to extract MBTiles layers for {}: {}", id, e);
-                    eprintln!("  File path: {}", full_path.display());
-                    eprintln!("  Tile format: {}", format);
-                    return Ok(Json(models::FileSchemaResponse { layers: vec![] }));
-                }
-            }
-        } else {
-            return Ok(Json(models::FileSchemaResponse { layers: vec![] }));
-        }
-    }
-
-    // Handle regular datasets (GeoJSON, Shapefile, etc.)
-    let mut cols_stmt = conn
-        .prepare(
-            "SELECT original_name, mvt_type\n         FROM dataset_columns\n         WHERE source_id = ?\n         ORDER BY ordinal",
-        )
-        .map_err(internal_error)?;
-
-    let cols_iter = cols_stmt
-        .query_map(duckdb::params![&id], |row| {
-            let original_name: String = row.get(0)?;
-            let mvt_type: String = row.get(1)?;
-            Ok((original_name, mvt_type))
-        })
-        .map_err(internal_error)?;
-
-    let mut fields = Vec::new();
-    for c in cols_iter {
-        let (name, r#type) = c.map_err(internal_error)?;
-        fields.push(models::FieldInfo { name, r#type });
-    }
-
-    drop(conn);
-
-    let default_layer = models::LayerInfo {
-        id: "default".to_string(),
-        description: None,
-        fields,
-    };
-    Ok(Json(models::FileSchemaResponse {
-        layers: vec![default_layer],
-    }))
-}
-
-fn validate_tile_coords(z: i32, x: i32, y: i32) -> Result<(), (StatusCode, Json<ErrorResponse>)> {
-    // Practical cap. This is plenty for web maps and keeps bounds math simple.
-    const MAX_Z: i32 = 22;
-
-    if z < 0 || x < 0 || y < 0 || z > MAX_Z {
-        return Err(bad_request("Invalid tile coordinates"));
-    }
-
-    let max_xy: i32 = 1_i32 << z;
-    if x >= max_xy || y >= max_xy {
-        return Err(bad_request("Invalid tile coordinates"));
-    }
-
-    Ok(())
-}
-
-async fn upload_file(
-    State(state): State<AppState>,
-    mut multipart: Multipart,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let mut field = loop {
-        let next = multipart.next_field().await.map_err(|e| {
-            let message = format!("Invalid multipart form: {e}");
-            bad_request(&message)
-        })?;
-        match next {
-            Some(field) if field.name() == Some("file") => break field,
-            Some(_) => continue,
-            None => return Err(bad_request("No file uploaded")),
-        }
-    };
-
-    let original_name = field
-        .file_name()
-        .map(|name| name.to_string())
-        .ok_or_else(|| bad_request("Missing file name"))?;
-    let safe_name = Path::new(&original_name)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .ok_or_else(|| bad_request("Invalid file name"))?
-        .to_string();
-
-    let ext = Path::new(&safe_name)
-        .extension()
-        .and_then(|ext| ext.to_str())
-        .map(|ext| format!(".{}", ext.to_lowercase()))
-        .ok_or_else(|| bad_request("Unsupported file type. Use .zip, .geojson, .json, .geojsonl, .kml, .gpx, .topojson, .mbtiles, or .pmtiles"))?;
-
-    let file_type = match ext.as_str() {
-        ".zip" => "shapefile",
-        ".geojson" | ".json" => "geojson",
-        ".geojsonl" | ".geojsons" => "geojsonl",
-        ".kml" => "kml",
-        ".gpx" => "gpx",
-        ".topojson" => "topojson",
-        ".mbtiles" => "mbtiles",
-        ".pmtiles" => "pmtiles",
-        _ => return Err(bad_request(
-            "Unsupported file type. Use .zip, .geojson, .json, .geojsonl, .kml, .gpx, .topojson, .mbtiles, or .pmtiles",
-        )),
-    };
-
-    let upload_id = create_id();
-    let dir = state.upload_dir.join(&upload_id);
-    fs::create_dir_all(&dir).await.map_err(internal_error)?;
-
-    let file_path = dir.join(&safe_name);
-    let mut file = BufWriter::new(fs::File::create(&file_path).await.map_err(internal_error)?);
-
-    let mut size: u64 = 0;
-    while let Some(chunk) = field.chunk().await.map_err(internal_error)? {
-        size = size.saturating_add(chunk.len() as u64);
-        if size > state.max_size {
-            drop(file);
-            let _ = fs::remove_file(&file_path).await;
-            let message = format!("File too large (max {})", state.max_size_label);
-            return Err(payload_too_large(&message));
-        }
-        file.write_all(&chunk).await.map_err(internal_error)?;
-    }
-    file.flush().await.map_err(internal_error)?;
-    drop(file); // Explicitly close file to release lock
-
-    let base_name = Path::new(&safe_name)
-        .file_stem()
-        .and_then(|name| name.to_str())
-        .unwrap_or(&safe_name)
-        .to_string();
-
-    let validation = match file_type {
-        "shapefile" => validate_shapefile_zip(&file_path).await,
-        "geojson" => validate_geojson(&file_path).await,
-        "mbtiles" => mbtiles::validate_mbtiles_structure(&file_path),
-        "pmtiles" => Ok(()), // PMTiles files are stored as-is, validated by frontend
-        "geojsonl" | "kml" | "gpx" | "topojson" => Ok(()), // Trust GDAL to validate
-        _ => Ok(()),         // Unreachable due to earlier validation, but required for type safety
-    };
-
-    let uploaded_at = Utc::now().to_rfc3339();
-
-    // Calculate relative path for storage
-    let relative = file_path
-        .strip_prefix(std::env::current_dir().unwrap_or_else(|_| PathBuf::from(".")))
-        .unwrap_or(&file_path)
-        .to_path_buf();
-    let mut rel_string = relative.to_string_lossy().replace('\\', "/");
-    if !rel_string.starts_with('.') {
-        rel_string = format!("./{rel_string}");
-    }
-
-    let conn = state.db.lock().await;
-
-    if let Err(message) = validation {
-        let size_i64 = size as i64;
-        conn.execute(
-            "INSERT INTO files (id, name, type, size, uploaded_at, status, crs, path, table_name, error, is_public)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11)",
-            duckdb::params![
-                &upload_id,
-                &base_name,
-                file_type,
-                size_i64,
-                &uploaded_at,
-                "failed",
-                &None::<String>,
-                &rel_string,
-                &None::<String>,
-                &Some(message.clone()),
-                false,
-            ],
-        )
-        .map_err(internal_error)?;
-
-        drop(conn);
-        return Err(bad_request(&message));
-    }
-
-    let size_i64 = size as i64;
-    conn.execute(
-        "INSERT INTO files (id, name, type, size, uploaded_at, status, crs, path, table_name, error, is_public)
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11)",
-        duckdb::params![
-            &upload_id,
-            &base_name,
-            file_type,
-            size_i64,
-            &uploaded_at,
-            "uploaded",
-            &None::<String>,
-            &rel_string,
-            &None::<String>,
-            &None::<String>,
-            false,
-        ],
-    )
-    .map_err(internal_error)?;
-
-    drop(conn);
-
-    let db = state.db.clone();
-    let upload_id_clone = upload_id.clone();
-    let file_path_clone = file_path.clone();
-    let file_type_clone = file_type.to_string();
-    tokio::spawn(async move {
-        // Set status to processing
-        {
-            let conn = db.lock().await;
-            let _ = conn.execute(
-                "UPDATE files SET status = 'processing' WHERE id = ?",
-                duckdb::params![upload_id_clone],
-            );
-        }
-
-        let result = match file_type_clone.as_str() {
-            "mbtiles" => import_mbtiles(&db, &upload_id_clone, &file_path_clone).await,
-            "pmtiles" => {
-                // PMTiles files are stored as-is, just mark as ready
-                let conn = db.lock().await;
-                let _ = conn.execute(
-                    "UPDATE files SET status = 'ready', tile_source = 'pmtiles' WHERE id = ?",
-                    duckdb::params![upload_id_clone],
-                );
-                Ok(())
-            }
-            _ => import_spatial_data(&db, &upload_id_clone, &file_path_clone).await,
-        };
-
-        match result {
-            Ok(_) => {
-                println!("Successfully imported spatial data for {}", upload_id_clone);
-                let conn = db.lock().await;
-                let _ = conn.execute(
-                    "UPDATE files SET status = 'ready' WHERE id = ?",
-                    duckdb::params![upload_id_clone],
-                );
-            }
-            Err(e) => {
-                eprintln!(
-                    "Failed to import spatial data for {}: {}",
-                    upload_id_clone, e
-                );
-                // Update status to failed
-                let conn = db.lock().await;
-                let _ = conn.execute(
-                    "UPDATE files SET status = 'failed', error = ? WHERE id = ?",
-                    duckdb::params![e, upload_id_clone],
-                );
-            }
-        }
-    });
-
-    let meta = FileItem {
-        id: upload_id,
-        name: base_name,
-        file_type: file_type.to_string(),
-        size,
-        uploaded_at,
-        status: "uploaded".to_string(),
-        crs: None,
-        path: rel_string,
-        table_name: None,
-        error: None,
-        is_public: Some(false),
-        public_slug: None,
-    };
-
-    Ok((StatusCode::CREATED, Json(meta)))
-}
-
-async fn health_check() -> impl IntoResponse {
-    (StatusCode::OK, Json(serde_json::json!({ "status": "ok" })))
-}
-
-async fn check_is_initialized(State(state): State<AppState>) -> impl IntoResponse {
-    let conn = state.db.lock().await;
-    match is_initialized(&conn) {
-        Ok(initialized) => (
-            StatusCode::OK,
-            Json(serde_json::json!({ "initialized": initialized })),
-        )
-            .into_response(),
-        Err(e) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(serde_json::json!({ "error": format!("Failed to check initialization status: {}", e) })),
-        )
-            .into_response(),
-    }
-}
-
-async fn publish_file(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-    Json(req): Json<PublishRequest>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let slug = match req.slug {
-        Some(s) => validate_slug(&s).map_err(|e| bad_request(&e))?,
-        None => validate_slug(&id).map_err(|e| bad_request(&e))?,
-    };
-
-    // Use transaction to ensure atomicity: insert into published_files first (enforces uniqueness),
-    // then update files table. This eliminates race conditions for concurrent publish requests.
-    conn.execute_batch("BEGIN TRANSACTION")
-        .map_err(internal_error)?;
-
-    // Check file status within transaction to provide better error messages
-    let (status, _name, tile_source): (String, String, String) = conn
-        .query_row(
-            "SELECT status, name, COALESCE(tile_source, 'duckdb') FROM files WHERE id = ?",
-            duckdb::params![&id],
-            |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "File not found".to_string(),
-                }),
-            )
-        })?;
-
-    if status != "ready" {
-        conn.execute_batch("ROLLBACK").map_err(internal_error)?;
-        drop(conn);
-        return Err((
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: format!("File is not ready for publishing (status: {})", status),
-            }),
-        ));
-    }
-
-    let insert_result = conn.execute(
-        "INSERT INTO published_files (file_id, slug, tile_source) VALUES (?, ?, ?)",
-        duckdb::params![&id, &slug, &tile_source],
-    );
-
-    let publish_result: Result<(), String> = match insert_result {
-        Ok(_) => conn
-            .execute(
-                "UPDATE files SET is_public = TRUE WHERE id = ?",
-                duckdb::params![&id],
-            )
-            .map(|_| ())
-            .map_err(|e| e.to_string()),
-        Err(e) => {
-            let err_msg = e.to_string();
-            // Check for PRIMARY KEY constraint on file_id (file already published) vs UNIQUE constraint on slug
-            // More specific detection: PRIMARY KEY on file_id, or constraint error mentioning file
-            let is_file_already_published = err_msg.contains("PRIMARY KEY")
-                || (err_msg.contains("Constraint Error") && err_msg.contains("file_id"));
-
-            if is_file_already_published {
-                // Immediately rollback the failed transaction
-                // After ROLLBACK, the connection returns to autocommit mode, allowing us to query
-                // without an active transaction. This is safe because:
-                // - The INSERT failed, so the transaction is aborted
-                // - We need to query for the existing slug to provide a helpful error message
-                // - The subsequent query runs in autocommit mode and does not affect database state
-                conn.execute_batch("ROLLBACK").map_err(internal_error)?;
-
-                // Query for existing slug (connection now in autocommit mode)
-                // We use .ok() here to convert "not found" to None. Any other error is also
-                // converted to None, which is acceptable given:
-                // - PRIMARY KEY constraint just confirmed the file exists
-                // - Query is simple and likely to succeed
-                // - If query fails, we return a generic but accurate error message
-                let existing_slug: Option<String> = conn
-                    .query_row(
-                        "SELECT slug FROM published_files WHERE file_id = ?",
-                        duckdb::params![&id],
-                        |row| row.get(0),
-                    )
-                    .ok();
-
-                drop(conn);
-
-                let error_msg = if let Some(existing) = existing_slug {
-                    format!(
-                        "File already published with slug '{existing}'. Unpublish first to change slug."
-                    )
-                } else {
-                    "File already published. Unpublish first to change slug.".to_string()
-                };
-
-                // Return early with the error (skip the outer match's ROLLBACK)
-                return Err((
-                    StatusCode::CONFLICT,
-                    Json(ErrorResponse { error: error_msg }),
-                ));
-            } else if err_msg.contains("UNIQUE")
-                || (err_msg.contains("slug") && err_msg.contains("unique"))
-            {
-                Err("Slug already in use".to_string())
-            } else {
-                Err(err_msg)
-            }
-        }
-    };
-
-    match publish_result {
-        Ok(()) => {
-            conn.execute_batch("COMMIT").map_err(internal_error)?;
-            drop(conn);
-            let url = if tile_source == "pmtiles" {
-                format!("/tiles/{slug}")
-            } else {
-                format!("/tiles/{slug}/{{z}}/{{x}}/{{y}}")
-            };
-            Ok(Json(PublishResponse {
-                url,
-                slug,
-                is_public: true,
-            }))
-        }
-        Err(err_msg) => {
-            conn.execute_batch("ROLLBACK").map_err(internal_error)?;
-            drop(conn);
-            Err(bad_request(&err_msg))
-        }
-    }
-}
-
-async fn unpublish_file(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    // Use transaction to ensure atomicity: delete from published_files and update files table
-    conn.execute_batch("BEGIN TRANSACTION")
-        .map_err(internal_error)?;
-
-    // Delete from published_files and verify file is actually published (is_public=TRUE)
-    // This ensures we don't leave orphaned published_files entries if files.is_public is FALSE
-    let rows_affected = conn
-        .execute(
-            "DELETE FROM published_files 
-            WHERE file_id = ? AND file_id IN (SELECT id FROM files WHERE is_public = TRUE)",
-            duckdb::params![&id],
-        )
-        .map_err(internal_error)?;
-
-    if rows_affected == 0 {
-        conn.execute_batch("ROLLBACK").map_err(internal_error)?;
-        drop(conn);
-        return Err((
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse {
-                error: "File not published".to_string(),
-            }),
-        ));
-    }
-
-    let update_result = conn
-        .execute(
-            "UPDATE files SET is_public = FALSE WHERE id = ?",
-            duckdb::params![&id],
-        )
-        .map_err(|e| e.to_string());
-
-    match update_result {
-        Ok(_) => {
-            conn.execute_batch("COMMIT").map_err(internal_error)?;
-            drop(conn);
-            Ok(Json(serde_json::json!({ "message": "File unpublished" })))
-        }
-        Err(err_msg) => {
-            conn.execute_batch("ROLLBACK").map_err(internal_error)?;
-            drop(conn);
-            Err(internal_error(err_msg.as_str()))
-        }
-    }
-}
-
-async fn get_public_url(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let result: Option<(String, String)> = conn
-        .query_row(
-            "SELECT pf.slug, pf.published_at FROM published_files pf JOIN files f ON pf.file_id = f.id WHERE f.id = ? AND f.is_public = TRUE",
-            duckdb::params![&id],
-            |row| Ok((row.get(0)?, row.get(1)?)),
-        )
-        .ok();
-
-    drop(conn);
-
-    match result {
-        Some((slug, _published_at)) => Ok(Json(PublicTileUrl {
-            slug: slug.clone(),
-            url: format!("/tiles/{slug}/{{z}}/{{x}}/{{y}}"),
-        })),
-        None => Err((
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse {
-                error: "File not published".to_string(),
-            }),
-        )),
-    }
-}
-
-async fn get_public_tile(
-    State(state): State<AppState>,
-    AxumPath((slug, z, x, y)): AxumPath<(String, i32, i32, i32)>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    validate_tile_coords(z, x, y)?;
-
-    let conn = state.db.lock().await;
-
-    // Step 1: Get file_id from published_files using slug (enforces uniqueness)
-    let file_id: String = conn
-        .query_row(
-            "SELECT file_id FROM published_files WHERE slug = ?",
-            duckdb::params![&slug],
-            |row| row.get(0),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "Public tile not found".to_string(),
-                }),
-            )
-        })?;
-
-    // Step 2: Get file metadata from files table, verifying is_public flag
-    let (crs, status, table_name, tile_format, file_path): (
-        Option<String>,
-        String,
-        Option<String>,
-        Option<String>,
-        String,
-    ) = conn
-        .query_row(
-            "SELECT crs, status, table_name, tile_format, path FROM files WHERE id = ? AND is_public = TRUE",
-            duckdb::params![&file_id],
-            |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?, row.get(4)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "File not found".to_string(),
-                }),
-            )
-        })?;
-
-    if status != "ready" {
-        return Err((
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready".to_string(),
-            }),
-        ));
-    }
-
-    // MBTiles branch
-    if let Some(format) = tile_format {
-        let full_path = mbtiles::resolve_mbtiles_path(&file_path);
-        drop(conn); // Release lock before async operation
-        match mbtiles::get_tile_from_mbtiles(&full_path, z, x, y).await {
-            Ok(Some(data)) => {
-                let ct = match format.as_str() {
-                    "mvt" => "application/vnd.mapbox-vector-tile",
-                    "png" => "image/png",
-                    _ => "application/octet-stream",
-                };
-                return Ok((
-                    [
-                        (header::CONTENT_TYPE, ct),
-                        (header::CACHE_CONTROL, "public, max-age=300"),
-                    ],
-                    data,
-                )
-                    .into_response());
-            }
-            Ok(None) => {
-                return Ok(StatusCode::NO_CONTENT.into_response());
-            }
-            Err(e) => {
-                return Err(internal_error(format!("Failed to read MBTiles: {}", e)));
-            }
-        }
-    }
-
-    // Dynamic generation branch (existing logic)
-    let table_name = table_name.ok_or_else(|| {
-        (
-            StatusCode::CONFLICT,
-            Json(ErrorResponse {
-                error: "File is not ready".to_string(),
-            }),
-        )
-    })?;
-
-    let source_crs = crs.as_deref().unwrap_or("EPSG:4326");
-
-    let select_sql =
-        build_mvt_select_sql(&conn, &file_id, &table_name, source_crs).map_err(internal_error)?;
-
-    let mvt_blob: Option<Vec<u8>> =
-        match conn.query_row(&select_sql, duckdb::params![z, x, y, z, x, y], |row| {
-            row.get(0)
-        }) {
-            Ok(blob) => Some(blob),
-            Err(e) => {
-                eprintln!("Tile Error (z={z}, x={x}, y={y}): {:?}", e);
-                return Err(internal_error(format!("Tile generation failed: {}", e)));
-            }
-        };
-
-    match mvt_blob {
-        Some(blob) if !blob.is_empty() => Ok((
-            [
-                (header::CONTENT_TYPE, "application/vnd.mapbox-vector-tile"),
-                (header::CACHE_CONTROL, "public, max-age=300"),
-            ],
-            blob,
-        )
-            .into_response()),
-        _ => Ok((
-            [
-                (header::CONTENT_TYPE, "application/vnd.mapbox-vector-tile"),
-                (header::CACHE_CONTROL, "public, max-age=300"),
-            ],
-            Vec::new(),
-        )
-            .into_response()),
-    }
-}
-
-async fn get_public_pmtiles(
-    State(state): State<AppState>,
-    AxumPath(slug): AxumPath<String>,
-    headers: HeaderMap,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let (_file_id, file_path, tile_source): (String, String, String) = conn
-        .query_row(
-            "SELECT f.id, f.path, COALESCE(pf.tile_source, f.tile_source, 'duckdb')
-             FROM files f
-             JOIN published_files pf ON f.id = pf.file_id
-             WHERE pf.slug = ? AND f.is_public = TRUE",
-            duckdb::params![&slug],
-            |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "Public tile not found".to_string(),
-                }),
-            )
-        })?;
-
-    if tile_source != "pmtiles" {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse {
-                error: "This endpoint is only for PMTiles files".to_string(),
-            }),
-        ));
-    }
-
-    let file_path = state
-        .upload_dir
-        .join(file_path.strip_prefix("./").unwrap_or(&file_path));
-    let canonical_path = file_path.canonicalize().map_err(|_| {
-        (
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse {
-                error: "PMTiles file not found".to_string(),
-            }),
-        )
-    })?;
-
-    if !canonical_path.starts_with(&state.upload_dir_canonical) {
-        return Err((
-            StatusCode::FORBIDDEN,
-            Json(ErrorResponse {
-                error: "Access denied".to_string(),
-            }),
-        ));
-    }
-
-    let file = fs::File::open(&canonical_path).await.map_err(|e| {
-        (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse {
-                error: format!("Failed to open file: {}", e),
-            }),
-        )
-    })?;
-
-    let metadata = file.metadata().await.map_err(|e| {
-        (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse {
-                error: format!("Failed to read metadata: {}", e),
-            }),
-        )
-    })?;
-    let file_size = metadata.len();
-
-    if let Some(range_header) = headers.get(header::RANGE) {
-        let range_str = range_header.to_str().map_err(|_| {
-            (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse {
-                    error: "Invalid Range header".to_string(),
-                }),
-            )
-        })?;
-
-        if !range_str.starts_with("bytes=") {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse {
-                    error: "Invalid Range header format".to_string(),
-                }),
-            ));
-        }
-
-        let range_spec = &range_str[6..];
-        let parts: Vec<&str> = range_spec.split('-').collect();
-        if parts.len() != 2 {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse {
-                    error: "Invalid Range header format".to_string(),
-                }),
-            ));
-        }
-
-        let start: u64 = if parts[0].is_empty() {
-            0
-        } else {
-            parts[0].parse().map_err(|_| {
-                (
-                    StatusCode::BAD_REQUEST,
-                    Json(ErrorResponse {
-                        error: "Invalid Range start".to_string(),
-                    }),
-                )
-            })?
-        };
-
-        let end: u64 = if parts[1].is_empty() {
-            file_size - 1
-        } else {
-            parts[1].parse().map_err(|_| {
-                (
-                    StatusCode::BAD_REQUEST,
-                    Json(ErrorResponse {
-                        error: "Invalid Range end".to_string(),
-                    }),
-                )
-            })?
-        };
-
-        if start > end || start >= file_size {
-            return Err((
-                StatusCode::RANGE_NOT_SATISFIABLE,
-                Json(ErrorResponse {
-                    error: "Range not satisfiable".to_string(),
-                }),
-            ));
-        }
-
-        let actual_end = end.min(file_size - 1);
-        let content_length = actual_end - start + 1;
-
-        use tokio::io::{AsyncReadExt, AsyncSeekExt};
-        let mut file = file;
-        file.seek(std::io::SeekFrom::Start(start))
-            .await
-            .map_err(|e| {
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ErrorResponse {
-                        error: format!("Seek failed: {}", e),
-                    }),
-                )
-            })?;
-
-        let mut buffer = vec![0u8; content_length as usize];
-        file.read_exact(&mut buffer).await.map_err(|e| {
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse {
-                    error: format!("Read failed: {}", e),
-                }),
-            )
-        })?;
-
-        Ok((
-            StatusCode::PARTIAL_CONTENT,
-            [
-                (header::CONTENT_TYPE, "application/octet-stream"),
-                (header::CONTENT_LENGTH, content_length.to_string().as_str()),
-                (
-                    header::CONTENT_RANGE,
-                    format!("bytes {}-{}/{}", start, actual_end, file_size).as_str(),
-                ),
-                (header::ACCEPT_RANGES, "bytes"),
-                (header::CACHE_CONTROL, "public, max-age=31536000, immutable"),
-            ],
-            buffer,
-        )
-            .into_response())
-    } else {
-        use tokio::io::AsyncReadExt;
-        let mut file = file;
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).await.map_err(|e| {
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse {
-                    error: format!("Read failed: {}", e),
-                }),
-            )
-        })?;
-
-        Ok((
-            StatusCode::OK,
-            [
-                (header::CONTENT_TYPE, "application/octet-stream"),
-                (header::CONTENT_LENGTH, file_size.to_string().as_str()),
-                (header::ACCEPT_RANGES, "bytes"),
-                (header::CACHE_CONTROL, "public, max-age=31536000, immutable"),
-            ],
-            buffer,
-        )
-            .into_response())
-    }
-}
-
-async fn head_public_pmtiles(
-    State(state): State<AppState>,
-    AxumPath(slug): AxumPath<String>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let (file_path, tile_source): (String, String) = conn
-        .query_row(
-            "SELECT f.path, COALESCE(pf.tile_source, f.tile_source, 'duckdb')
-             FROM files f
-             JOIN published_files pf ON f.id = pf.file_id
-             WHERE pf.slug = ? AND f.is_public = TRUE",
-            duckdb::params![&slug],
-            |row| Ok((row.get(0)?, row.get(1)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "Public tile not found".to_string(),
-                }),
-            )
-        })?;
-
-    if tile_source != "pmtiles" {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse {
-                error: "This endpoint is only for PMTiles files".to_string(),
-            }),
-        ));
-    }
-
-    let file_path = state
-        .upload_dir
-        .join(file_path.strip_prefix("./").unwrap_or(&file_path));
-    let canonical_path = file_path.canonicalize().map_err(|_| {
-        (
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse {
-                error: "PMTiles file not found".to_string(),
-            }),
-        )
-    })?;
-
-    if !canonical_path.starts_with(&state.upload_dir_canonical) {
-        return Err((
-            StatusCode::FORBIDDEN,
-            Json(ErrorResponse {
-                error: "Access denied".to_string(),
-            }),
-        ));
-    }
-
-    let metadata = fs::metadata(&canonical_path).await.map_err(|e| {
-        (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse {
-                error: format!("Failed to read metadata: {}", e),
-            }),
-        )
-    })?;
-    let file_size = metadata.len();
-
-    Ok((
-        StatusCode::OK,
-        [
-            (header::CONTENT_TYPE, "application/octet-stream"),
-            (header::CONTENT_LENGTH, file_size.to_string().as_str()),
-            (header::ACCEPT_RANGES, "bytes"),
-            (header::CACHE_CONTROL, "public, max-age=31536000, immutable"),
-        ],
-    )
-        .into_response())
-}
-
-async fn get_public_tile_meta(
-    State(state): State<AppState>,
-    AxumPath(slug): AxumPath<String>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    let conn = state.db.lock().await;
-
-    let (name, tile_source): (String, String) = conn
-        .query_row(
-            "SELECT f.name, COALESCE(pf.tile_source, f.tile_source, 'duckdb')
-             FROM files f
-             JOIN published_files pf ON f.id = pf.file_id
-             WHERE pf.slug = ? AND f.is_public = TRUE",
-            duckdb::params![&slug],
-            |row| Ok((row.get(0)?, row.get(1)?)),
-        )
-        .map_err(|_| {
-            (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse {
-                    error: "Public tile not found".to_string(),
-                }),
-            )
-        })?;
-
-    let tile_url = if tile_source == "pmtiles" {
-        format!("/tiles/{}", slug)
-    } else {
-        format!("/tiles/{}/{{z}}/{{x}}/{{y}}", slug)
-    };
-    let viewer_url = format!("/tiles/{}", slug);
-
-    Ok(Json(PublicTileMeta {
-        slug: slug.clone(),
-        name,
-        tile_source,
-        tile_url,
-        viewer_url,
-    }))
-}
-
-fn validate_slug(slug: &str) -> Result<String, String> {
-    let slug = slug.trim().to_string();
-
-    if slug.is_empty() {
-        return Err("Slug cannot be empty".to_string());
-    }
-
-    if slug.len() > 100 {
-        return Err("Slug must be 100 characters or less".to_string());
-    }
-
-    if !slug
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || c == '-' || c == '_')
-    {
-        return Err("Slug can only contain letters, numbers, hyphens, and underscores".to_string());
-    }
-
-    Ok(slug)
-}
-
-fn create_id() -> String {
-    let mut bytes = [0u8; 3];
-    rand::thread_rng().fill_bytes(&mut bytes);
-    hex::encode(bytes)
-}
 
 #[cfg(test)]
 mod tests {
@@ -1605,7 +50,7 @@ mod tests {
     async fn setup_state(max_size: u64) -> (AppState, TempDir) {
         let temp_dir = TempDir::new().expect("temp dir");
         let upload_dir = temp_dir.path().join("uploads");
-        fs::create_dir_all(&upload_dir).await.ok();
+        tokio::fs::create_dir_all(&upload_dir).await.ok();
         let upload_dir_canonical = upload_dir
             .canonicalize()
             .unwrap_or_else(|_| upload_dir.clone());
@@ -1732,7 +177,7 @@ mod tests {
             .await
             .unwrap();
 
-        assert_eq!(response.status(), StatusCode::OK);
+        assert_eq!(response.status(), axum::http::StatusCode::OK);
         let items: Vec<FileItem> = response_json(response).await;
         assert_eq!(items.len(), 1);
         assert_eq!(items[0].name, "existing");
@@ -1746,18 +191,15 @@ mod tests {
             .lock()
             .expect("env lock");
 
-        // Default to false
         std::env::remove_var("COOKIE_SECURE");
         assert!(!read_cookie_secure());
 
-        // Explicitly set to false
         std::env::set_var("COOKIE_SECURE", "false");
         assert!(!read_cookie_secure());
 
         std::env::set_var("COOKIE_SECURE", "true");
         assert!(read_cookie_secure());
 
-        // Invalid value falls back to default false
         std::env::set_var("COOKIE_SECURE", "invalid");
         assert!(!read_cookie_secure());
